+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+- Remove drop() and dropWhile() from $List and $Iterable (use skip() and skipWhile() instead)
+- Add $it() as short form of $Iterable()
+- Add test cases
+
 ## 0.1.3
 - Add joinToString(), onEach() and shuffled to $Iterable